pyproject.toml
[build-system]
requires = ["maturin>=1.0,<2.0"]
build-backend = "maturin"

[project]
name = "happycode"
version = "0.1.0"
description = "Code-graph-aware AI software engineering toolkit"
requires-python = ">=3.9"
license = { text = "Apache-2.0" }
dependencies = [
    "rlms>=0.1.0",
    "litellm>=1.0.0",
    "tomli>=2.0; python_version < '3.11'",
]

[project.scripts]
happy-rlm = "happy_faster_code.orchestrator:main"
happy-launch = "happy_faster_code.launch:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "maturin>=1.0",
]

[tool.maturin]
manifest-path = "crates/happy-core/Cargo.toml"
features = ["python"]
python-source = "python"
module-name = "happy_faster_code.happy_core"
